requirements.txt
office365-rest-python-client==2.5.3
python-dotenv==1.0.0
pandas==2.1.4
openpyxl==3.1.2
